requirements.txt
Pillow>=11.3.0
tqdm>=4.67.1
InquirerPy>=0.3.4
global-land-mask>=1.0.0